refactor(router-cli): split TUI into focused widget modules
Break ui.rs (482 lines) into tui/widgets/ with one file per
concern: header, overview, stats, sparklines, status_codes,
footer, help. Remove section comments. ui.rs is now a thin
layout orchestrator.
