+++ router-cli/src/tui/widgets/header.rs
@@ -1,0 +1,21 @@
+use ratatui::layout::Rect;
+use ratatui::style::{Color, Modifier, Style};
+use ratatui::text::{Line, Span};
+use ratatui::widgets::Paragraph;
+use ratatui::Frame;
+
+pub fn draw(f: &mut Frame, area: Rect) {
+    let header = Line::from(vec![
+        Span::styled(
+            " Mini-Gateway Monitor ",
+            Style::default()
+                .fg(Color::Cyan)
+                .add_modifier(Modifier::BOLD),
+        ),
+        Span::styled(
+            "q:quit  tab:switch  1:gw  2:px  ?:help",
+            Style::default().fg(Color::DarkGray),
+        ),
+    ]);
+    f.render_widget(Paragraph::new(header), area);
+}
